NF: Classes to determine input type
