+++ psychopy/app/builder/dialogs/__init__.py
@@ -10,7 +10,6 @@
 
 from __future__ import absolute_import, division, print_function
 import sys
-from pathlib import Path
 
 from builtins import map
 from builtins import str
@@ -28,6 +27,7 @@ from .. import experiment
 from .. validators import NameValidator, CodeSnippetValidator
 from .dlgsConditions import DlgConditions
 from .dlgsCode import DlgCodeComponentProperties, CodeBox
+from . import paramCtrls
 from psychopy import data, logging
 from psychopy.localization import _translate
 from psychopy.tools import versionchooser as vc
@@ -122,8 +122,8 @@ class ParamCtrls(object):
             else:
                 sx, sy = 100, 200
             # set viewer small, then it SHOULD increase with wx.aui control
-            self.valueCtrl = wx.TextCtrl(parent, -1, value=str(param.val), pos=wx.DefaultPosition,
-                                         size=wx.Size(sx, sy), style=wx.TE_MULTILINE)
+            self.valueCtrl = paramCtrls.ExtendedStringCtrl(parent, val=str(param.val), fieldName=fieldName,
+                                         size=wx.Size(sx, sy))
             if fieldName == 'text':
                 self.valueCtrl.SetFocus()
         elif fieldName == 'Experiment info':
@@ -149,17 +149,17 @@ class ParamCtrls(object):
             self.valueCtrl.SetCheckedStrings(param.val)
         elif param.valType == 'bool':
             # only True or False - use a checkbox
-            self.valueCtrl = wx.CheckBox(parent,
+            self.valueCtrl = paramCtrls.BoolCtrl(parent,
                                          name=fieldName,
                                          size=wx.Size(self.valueWidth, -1))
             self.valueCtrl.SetValue(param.val)
         elif param.valType == 'fileList':
-            self.valueCtrl = FileListCtrl(parent,
+            self.valueCtrl = paramCtrls.FileListCtrl(parent,
                                           choices=param.val,
                                           size=wx.Size(self.valueWidth, 100),
                                           pathtype="rel")
         elif param.valType == 'table':
-            self.valueCtrl = TableCtrl(parent, val=param.val, fieldName=fieldName,
+            self.valueCtrl = paramCtrls.TableCtrl(parent, val=param.val, fieldName=fieldName,
                                        size=wx.Size(self.valueWidth, 16))
         elif len(param.allowedVals) > 1:
             # there are limited options - use a Choice control
@@ -511,11 +511,7 @@ class _BaseParamsDlg(wx.Dialog):
                       'Online':_translate('Online'),
                       'Testing':_translate('Testing'),
                       'Audio':_translate('Audio'),
-                      'Format':_translate('Format'),
-                      'Formatting':_translate('Formatting'),
-                      'Texture':_translate('Texture'),
-                      'Timing':_translate('Timing'),
-                      'Playback':_translate('Playback')}
+                      'Format':_translate('Format')}
         for categName in categNames:
             theseParams = categs[categName]
             page = wx.Panel(self.ctrls, -1)
@@ -1348,7 +1344,6 @@ class DlgLoopProperties(_BaseParamsDlg):
                 self.Bind(wx.EVT_BUTTON, self.onBrowseTrialsFile,
                           ctrls.browseCtrl)
                 ctrls.valueCtrl.Bind(wx.EVT_RIGHT_DOWN, self.viewConditions)
-                ctrls.valueCtrl.Bind(wx.EVT_TEXT, self.doValidate)
                 panelSizer.Add(ctrls.nameCtrl, [row, 0])
                 panelSizer.Add(ctrls.valueCtrl, [row, 1])
                 panelSizer.Add(ctrls.browseCtrl, [row, 2])
@@ -1940,138 +1935,6 @@ class DlgExperimentProperties(_BaseParamsDlg):
         else:
             self.OK = False
         return wx.ID_OK
-
-class FileListCtrl(wx.ListBox):
-    def __init__(self, parent, choices=[], size=None, pathtype="rel"):
-        wx.ListBox.__init__(self)
-        parent.Bind(wx.EVT_DROP_FILES, self.addItem)
-        if type(choices) == str:
-            choices = data.utils.listFromString(choices)
-        self.Create(id=wx.ID_ANY, parent=parent, choices=choices, size=size, style=wx.LB_EXTENDED | wx.LB_HSCROLL)
-        # can now get/set things like self.parent
-        self.app = parent.app
-        self.builderFrame = self.GetTopLevelParent().frame
-        self.addBtn = wx.Button(parent, -1, style=wx.BU_EXACTFIT, label="+")
-        self.addBtn.Bind(wx.EVT_BUTTON, self.addItem)
-        self.subBtn = wx.Button(parent, -1, style=wx.BU_EXACTFIT, label="-")
-        self.subBtn.Bind(wx.EVT_BUTTON, self.removeItem)
-
-        self._szr = wx.BoxSizer(wx.HORIZONTAL)
-        self.btns = wx.BoxSizer(wx.VERTICAL)
-        self.btns.AddMany((self.addBtn, self.subBtn))
-        self._szr.Add(self, proportion=1, flag=wx.EXPAND)
-        self._szr.Add(self.btns)
-
-    def addItem(self, event):
-        if event.GetEventObject() == self.addBtn:
-            _wld = "Any file (*.*)|*"
-            dlg = wx.FileDialog(self, message=_translate("Specify file ..."),
-                                style=wx.FD_OPEN | wx.FD_FILE_MUST_EXIST | wx.FD_MULTIPLE,
-                                wildcard=_translate(_wld))
-            if dlg.ShowModal() != wx.ID_OK:
-                return 0
-            fileList = dlg.GetPaths()
-        else:
-            fileList = event.GetFiles()
-        relPaths = []
-        expFile = self.builderFrame.filename
-        folder = Path(expFile).parent
-        for filename in fileList:
-            relPaths.append(
-                os.path.relpath(filename, folder))
-        self.InsertItems(relPaths, 0)
-
-    def removeItem(self, event):
-        i = self.GetSelections()
-        if isinstance(i, int):
-            i = [i]
-        items = [item for index, item in enumerate(self.Items)
-                 if index not in i]
-        self.SetItems(items)
-
-    def GetValue(self):
-        return self.Items
-
-class TableCtrl(wx.TextCtrl):
-    def __init__(self, parent,
-                 val="", fieldName="",
-                 size=wx.Size(-1, 24)):
-        # Create self
-        wx.TextCtrl.__init__(self)
-        self.Create(parent, -1, val, name=fieldName, size=size)
-        # Add sizer
-        self._szr = wx.BoxSizer(wx.HORIZONTAL)
-        self._szr.Add(self, border=5, flag=wx.EXPAND | wx.RIGHT)
-        # Add button to browse for file
-        fldr = parent.app.iconCache.getBitmap(name="folder", size=16, theme="light")
-        self.findBtn = wx.BitmapButton(parent, -1, size=wx.Size(24,24), bitmap=fldr)
-        self.findBtn.SetToolTip(_translate("Specify file ..."))
-        self.findBtn.Bind(wx.EVT_BUTTON, self.findFile)
-        self._szr.Add(self.findBtn)
-        # Add button to open in Excel
-        xl = parent.app.iconCache.getBitmap(name="filecsv", size=16, theme="light")
-        self.xlBtn = wx.BitmapButton(parent, -1, size=wx.Size(24,24), bitmap=xl)
-        self.xlBtn.SetToolTip(_translate("Open/create in your default table editor"))
-        self.xlBtn.Bind(wx.EVT_BUTTON, self.openExcel)
-        self._szr.Add(self.xlBtn)
-        # Link to Excel templates for certain contexts
-        cmpRoot = os.path.dirname(psychopy.experiment.components.__file__)
-        self.templates = {
-            'Form': os.path.join(cmpRoot, "form", "formItems.xltx")
-        }
-        # Configure validation
-        self.Bind(wx.EVT_TEXT, self.validateInput)
-        self.validExt = [".csv",".tsv",".txt",
-                         ".xl",".xlsx",".xlsm",".xlsb",".xlam",".xltx",".xltm",".xls",".xlt",
-                         ".htm",".html",".mht",".mhtml",
-                         ".xml",".xla",".xlm",
-                         ".odc",".ods",
-                         ".udl",".dsn",".mdb",".mde",".accdb",".accde",".dbc",".dbf",
-                         ".iqy",".dqy",".rqy",".oqy",
-                         ".cub",".atom",".atomsvc",
-                         ".prn",".slk",".dif"]
-
-    def validateInput(self, event):
-        """Check whether input is openable and valid"""
-        valid = False
-        file = self.GetValue()
-        # Is component type available?
-        if hasattr(self.GetTopLevelParent(), 'type'):
-            # Does this component have a default template?
-            if self.GetTopLevelParent().type in self.templates:
-                valid = True
-        # Has user entered a full filepath, but it is invalid?
-        if file and file not in self.validExt:
-            valid = False
-        # Is value a valid filepath?
-        if os.path.isfile(os.path.abspath(file)) and file.endswith(tuple(self.validExt)):
-            valid = True
-        # Set excel button accordingly
-        self.xlBtn.Enable(valid)
-
-    def openExcel(self, event):
-        """Either open the specified excel sheet, or make a new one from a template"""
-        file = self.GetValue()
-        if os.path.isfile(file) and file.endswith(tuple(self.validExt)):
-            os.startfile(file)
-        else:
-            dlg = wx.MessageDialog(self, _translate(
-                f"Once you have created and saved your table, please remember to add it to {self.Name}"),
-                             caption="Reminder")
-            dlg.ShowModal()
-            os.startfile(self.templates[self.GetTopLevelParent().type])
-
-    def findFile(self, event):
-        _wld = f"All Table Files({'*'+';*'.join(self.validExt)})|{'*'+';*'.join(self.validExt)}|All Files (*.*)|*.*"
-        dlg = wx.FileDialog(self, message=_translate("Specify file ..."),
-                            style=wx.FD_OPEN | wx.FD_FILE_MUST_EXIST,
-                            wildcard=_translate(_wld))
-        if dlg.ShowModal() != wx.ID_OK:
-            return 0
-        filename = dlg.GetPath()
-        relname = os.path.relpath(filename)
-        self.SetValue(relname)
-        self.validateInput(event)
 
 def _relpath(path, start='.'):
     """This code is based on os.path.relpath in the Python 2.6 distribution,
